Moved VmAccountInfo and other Vm classes into bpf_loader's mod.rs
