+++ sdk/program/src/account_info.rs
@@ -395,75 +395,6 @@ impl<'a> AsRef<AccountInfo<'a>> for AccountInfo<'a> {
     }
 }
 
-// The VmSlice class as defined in bpf_loader
-#[derive(Clone)]
-#[repr(C)]
-pub struct VmSlice<T> {
-    ptr: u64,
-    len: u64,
-    resource_type: PhantomData<T>,
-}
-
-impl<T> VmSlice<T> {
-    pub fn ptr(&self) -> u64 {
-        self.ptr
-    }
-    pub fn len(&self) -> u64 {
-        self.len
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.len == 0
-    }
-}
-
-// Structs to allow the AccountInfo translation to properly reference elements within
-// the 64-bit virtual address space even when built in 32-bit mode.
-#[derive(Clone)]
-#[repr(C)]
-pub struct VmNonNull<T>
-{
-    pub addr: u64,
-    resource_type: PhantomData<T>,
-}
-
-#[derive(Clone)]
-#[repr(C)]
-pub struct VmBoxOfRefCell<T>
-{
-    _strong_addr: u64,
-    _weak_addr: u64,
-    _borrow_flag: u64,
-    pub value: T,
-}
-
-/// Account information, in the virtual address space. Note: Since the addresses are u64,
-/// there is no lifetime on this struct, and the borrow checker cannot properly reason
-/// about references to the virtual memory via these addresses.
-#[derive(Clone)]
-#[repr(C)]
-pub struct VmAccountInfo<'a> {
-    /// Public key of the account (&'a Pubkey)
-    pub key: u64,
-    /// The address to the lamports in the account.  Modifiable by programs. (in `AccountInfo`: &'a mut u64)
-    pub lamports: VmNonNull<VmBoxOfRefCell<u64>>,
-    /// The data slice held in this account.  Modifiable by programs. (In `AccountInfo`: &'a mut [u8])
-    pub data: VmNonNull<VmBoxOfRefCell<VmSlice<u8>>>,
-    /// Program that owns this account (in `AccountInfo`: &'a Pubkey)
-    pub owner: u64,
-    /// The epoch at which this account will next owe rent
-    pub rent_epoch: Epoch,
-
-    /// Was the transaction signed by this account's public key?
-    pub is_signer: bool,
-    /// Is the account writable?
-    pub is_writable: bool,
-    /// This account's data contains a loaded program (and is now read-only)
-    pub executable: bool,
-
-    phantom: PhantomData<&'a u8>,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
